refactor: move error handlers to separate module
