+++ src/warehouse/errorhandlers.py
@@ -1,0 +1,37 @@
+# Copyright (c) 2018, Novo Nordisk Foundation Center for Biosustainability,
+# Technical University of Denmark.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Define Flask error handlers."""
+
+from flask import jsonify
+from werkzeug.exceptions import HTTPException
+
+
+def init_app(app):
+    # Add an error handler for webargs parser error, ensuring a JSON response
+    # including all error messages produced from the parser.
+    @app.errorhandler(422)
+    def handle_webargs_error(error):
+        response = jsonify(error.data['messages'])
+        response.status_code = error.code
+        return response
+
+    # Handle werkzeug HTTPExceptions (typically raised through `flask.abort`) by
+    # returning a JSON response including the error description.
+    @app.errorhandler(HTTPException)
+    def handle_error(error):
+        response = jsonify({'message': error.description})
+        response.status_code = error.code
+        return response
